Moving Files to a proper folder structure
