+++ Mastermind/Randomizer/IRandomizer.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Mastermind.Enums;
 
 namespace Mastermind.Randomizer
 {
